Import and explain govuk_elements
- Import elements at the top of application.scss
- Explain how styles for #content, #content h1 and .alpha-banner exist
within elements and where to find those
- Add a comment about how to continue in the easiest way possible - to
continue adding Sass at the bottom of application.scss, or to modify
govuk elements, or to re-arrange and those files as makes most sense to
the person using the prototype kit.
- This way, anyone copying and pasting snippets from govuk elements
into the prototype kit, will see that code just work. Which is a good
place to start.